requirements.txt
flask
flask-cors
flask-socketio
flask-sqlalchemy
eventlet
python-socketio
python-engineio
werkzeug
openai
python-dotenv